+++ Dozer/AppDelegate.swift
@@ -3,8 +3,6 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
 import Cocoa
-import Crashlytics
-import Fabric
 import MASShortcut
 import Sparkle
 import Defaults
@@ -13,12 +11,7 @@ import Preferences
 
 final class AppDelegate: NSObject, NSApplicationDelegate {
     func applicationDidFinishLaunching(_: Notification) {
-        #if !DEBUG
-            UserDefaults.standard.register(defaults: ["NSApplicationCrashOnExceptions": true])
-            Fabric.with([Crashlytics.self])
-        #endif
-
-        MASShortcutBinder.shared()?.bindShortcut(withDefaultsKey: UserDefaultKeys.Shortcuts.ToggleMenuItems) { [unowned self] in
+        MASShortcutBinder.shared()?.bindShortcut(withDefaultsKey: UserDefaultKeys.Shortcuts.ToggleMenuItems) { () in
             DozerIcons.shared.toggle()
         }
 
@@ -27,6 +20,8 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
         if DozerIcons.shared.hideStatusBarIconsAtLaunch {
             DozerIcons.shared.hide()
         }
+
+        _ = DozerIcons.toggleDockIcon(showIcon: false)
     }
 
     // Show all Dozer icons when opening Dozer from Finder etc.