+++ Dozer/ViewControllers/GeneralVC.swift
@@ -10,13 +10,11 @@ import Sparkle
 import Defaults
 
 final class General: NSViewController, PreferencePane {
-    let preferencePaneIdentifier = PreferencePaneIdentifier.general
+    let preferencePaneIdentifier = Preferences.PaneIdentifier.general
     let preferencePaneTitle: String = "General"
     let toolbarItemIcon = NSImage(named: NSImage.preferencesGeneralName)!
 
-    override var nibName: NSNib.Name? {
-        return "General"
-    }
+    override var nibName: NSNib.Name? { "General" }
 
     fileprivate var userShortCut: MASShortcut!
 
@@ -24,8 +22,12 @@ final class General: NSViewController, PreferencePane {
     @IBOutlet private var CheckForUpdatesCheckbox: NSButton!
     @IBOutlet private var HideStatusBarIconsAtLaunchCheckbox: NSButton!
     @IBOutlet private var HideStatusBarIconsAfterDelayCheckbox: NSButton!
+    @IBOutlet private var HideStatusBarIconsSecondsPopUpButton: NSPopUpButton!
     @IBOutlet private var HideBothDozerIconsCheckbox: NSButton!
     @IBOutlet private var EnableRemoveDozerIconCheckbox: NSButton!
+    @IBOutlet private var ShowIconAndMenuCheckbox: NSButton!
+    @IBOutlet private var FontSizePopUpButton: NSPopUpButton!
+    @IBOutlet private var ButtonPaddingPopUpButton: NSPopUpButton!
     @IBOutlet private var ToggleMenuItemsView: MASShortcutView!
 
     override func viewDidLoad() {
@@ -40,16 +42,22 @@ final class General: NSViewController, PreferencePane {
             CheckForUpdatesCheckbox.isChecked = false
         }
 
-        HideStatusBarIconsAtLaunchCheckbox.isChecked = defaults[.hideAtLaunchEnabled]
-        HideStatusBarIconsAfterDelayCheckbox.isChecked = defaults[.hideAfterDelayEnabled]
-        HideBothDozerIconsCheckbox.isChecked = defaults[.noIconMode]
-        EnableRemoveDozerIconCheckbox.isChecked = defaults[.removeDozerIconEnabled]
+        HideStatusBarIconsAtLaunchCheckbox.isChecked = Defaults[.hideAtLaunchEnabled]
+        HideStatusBarIconsAfterDelayCheckbox.isChecked = Defaults[.hideAfterDelayEnabled]
+        HideBothDozerIconsCheckbox.isChecked = Defaults[.noIconMode]
+        EnableRemoveDozerIconCheckbox.isChecked = Defaults[.removeDozerIconEnabled]
+        ShowIconAndMenuCheckbox.isChecked = Defaults[.showIconAndMenuEnabled]
+        HideStatusBarIconsSecondsPopUpButton.selectItem(withTitle: "\(Int(Defaults[.hideAfterDelay])) seconds")
+        FontSizePopUpButton.selectItem(withTitle: "\(Int(Defaults[.iconSize])) px")
+        ButtonPaddingPopUpButton.selectItem(withTitle: "\(Int(Defaults[.buttonPadding])) px")
 
         ToggleMenuItemsView.associatedUserDefaultsKey = UserDefaultKeys.Shortcuts.ToggleMenuItems
         view.addSubview(ToggleMenuItemsView)
+        configureEnabledNoIconCheckbox()
 
         ToggleMenuItemsView.shortcutValueChange = { _ -> Void in
             self.userShortCut = self.ToggleMenuItemsView.shortcutValue
+            self.configureEnabledNoIconCheckbox()
         }
     }
 
@@ -64,7 +72,7 @@ final class General: NSViewController, PreferencePane {
         }
         SUUpdater.shared()!.automaticallyChecksForUpdates = CheckForUpdatesCheckbox.isChecked
     }
-    
+
     @IBAction private func hideStatusBarIconsAtLaunchClicked(_ sender: NSButton) {
         DozerIcons.shared.hideStatusBarIconsAtLaunch = HideStatusBarIconsAtLaunchCheckbox.isChecked
     }
@@ -73,11 +81,39 @@ final class General: NSViewController, PreferencePane {
         DozerIcons.shared.hideStatusBarIconsAfterDelay = HideStatusBarIconsAfterDelayCheckbox.isChecked
     }
 
+    @IBAction private func hideStatusBarIconsSecondsUpdated(_ sender: NSPopUpButton) {
+        Defaults[.hideAfterDelay] = TimeInterval(HideStatusBarIconsSecondsPopUpButton.selectedTag())
+        DozerIcons.shared.resetTimer()
+    }
+
     @IBAction private func hideBothDozerIconsClicked(_ sender: NSButton) {
         DozerIcons.shared.hideBothDozerIcons = HideBothDozerIconsCheckbox.isChecked
     }
 
+    @IBAction private func showIconAndMenuClicked(_ sender: NSButton) {
+        DozerIcons.shared.enableIconAndMenu = ShowIconAndMenuCheckbox.isChecked
+    }
+
+    @IBAction private func fontSizeChanged(_ sender: NSPopUpButton) {
+        DozerIcons.shared.iconFontSize = FontSizePopUpButton.selectedTag()
+    }
+
+    @IBAction private func buttonPaddingChanged(_ sender: NSPopUpButton) {
+        DozerIcons.shared.buttonPadding = CGFloat(ButtonPaddingPopUpButton.selectedTag())
+    }
+
     @IBAction private func enableRemoveDozerIconClicked(_ sender: NSButton) {
         DozerIcons.shared.enableRemoveDozerIcon = EnableRemoveDozerIconCheckbox.isChecked
+    }
+
+    /// disables the noIcon-checkbox if no shortcut is set and keeps track whether shortcut is set
+    private func configureEnabledNoIconCheckbox() {
+        if ToggleMenuItemsView.shortcutValue == nil {
+            HideBothDozerIconsCheckbox.isEnabled = false
+            Defaults[.isShortcutSet] = false
+        } else {
+            HideBothDozerIconsCheckbox.isEnabled = true
+            Defaults[.isShortcutSet] = true
+        }
     }
 }